Remove useMenuPopup hook
