+++ packages/react/src/menu/popup/MenuPopup.tsx
@@ -1,8 +1,8 @@
 'use client';
 import * as React from 'react';
 import { FloatingFocusManager, useFloatingTree } from '@floating-ui/react';
-import { useMenuPopup } from './useMenuPopup';
 import { useMenuRootContext } from '../root/MenuRootContext';
+import type { MenuRoot } from '../root/MenuRoot';
 import { useMenuPositionerContext } from '../positioner/MenuPositionerContext';
 import { useComponentRenderer } from '../../utils/useComponentRenderer';
 import { useForkRef } from '../../utils/useForkRef';
@@ -61,10 +61,20 @@ export const MenuPopup = React.forwardRef(function MenuPopup(
 
   const { events: menuEvents } = useFloatingTree()!;
 
-  useMenuPopup({
-    setOpen,
-    menuEvents,
-  });
+  React.useEffect(() => {
+    function handleClose(event: {
+      domEvent: Event | undefined;
+      reason: MenuRoot.OpenChangeReason | undefined;
+    }) {
+      setOpen(false, event.domEvent, event.reason);
+    }
+
+    menuEvents.on('close', handleClose);
+
+    return () => {
+      menuEvents.off('close', handleClose);
+    };
+  }, [menuEvents, setOpen]);
 
   const mergedRef = useForkRef(forwardedRef, popupRef);
